+++ src/background/background.ts
@@ -1,7 +1,11 @@
 import browser from "webextension-polyfill";
+import onActivated from "./handlers/onActivated";
 import onBeforeRequest from "./handlers/onBeforeRequest";
 import onBeforeSendHeaders from "./handlers/onBeforeSendHeaders";
 import onHeadersReceived from "./handlers/onHeadersReceived";
+
+// Update browser action icon.
+browser.tabs.onActivated.addListener(onActivated);
 
 // Redirect the HLS master manifest request to TTV LOL's API.
 browser.webRequest.onBeforeRequest.addListener(
@@ -28,25 +32,3 @@ browser.webRequest.onHeadersReceived.addListener(
   { urls: ["https://api.ttv.lol/playlist/*", "https://api.ttv.lol/vod/*"] },
   ["blocking"]
 );
-
-// Update browser action icon.
-browser.tabs.onActivated.addListener(async () => {
-  const icon = await getBrowserActionIcon();
-  browser.browserAction.setIcon(icon);
-});
-
-async function getBrowserActionIcon(): Promise<browser.Action.SetIconDetailsType> {
-  const ICON_GRAY_PATH = new URL("../assets/icon-gray.png", import.meta.url);
-  const ICON_PATH = new URL("../assets/icon.png", import.meta.url);
-
-  const tabs = await browser.tabs.query({ active: true, currentWindow: true });
-  const activeTab = tabs[0];
-  if (!activeTab || !activeTab.url) return { path: ICON_GRAY_PATH.href };
-  try {
-    const url = new URL(activeTab.url);
-    if (url.hostname === "www.twitch.tv") {
-      return { path: ICON_PATH.href };
-    }
-  } catch {}
-  return { path: ICON_GRAY_PATH.href };
-}
